Extract connection helpers from __init__
