+++ cmk/gui/userdb/__init__.py
@@ -3,32 +3,7 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
-# TODO: Rework connection management and multiplexing
-from __future__ import annotations
-
-import ast
-import shutil
-import time
-import traceback
-from collections.abc import Callable, Sequence
-from contextlib import suppress
-from datetime import datetime
-from logging import Logger
-from pathlib import Path
-from typing import Any, Literal
-
-import cmk.utils.paths
-
 import cmk.gui.utils as utils
-from cmk.gui.background_job import (
-    BackgroundJob,
-    BackgroundJobAlreadyRunning,
-    BackgroundJobRegistry,
-    BackgroundProcessInterface,
-    InitialStatusArgs,
-)
-from cmk.gui.config import active_config
-from cmk.gui.i18n import _
 from cmk.gui.plugins.userdb.utils import (
     active_connections,
     ConnectorType,
@@ -63,17 +38,6 @@ from cmk.gui.userdb.store import (
     save_users,
     write_contacts_and_users_file,
 )
-from cmk.gui.utils.urls import makeuri_contextless
-from cmk.gui.valuespec import (
-    DEF_VALUE,
-    DropdownChoice,
-    TextInput,
-    Transform,
-    ValueSpec,
-    ValueSpecDefault,
-    ValueSpecHelp,
-    ValueSpecText,
-)
 
 from ._check_credentials import check_credentials as check_credentials
 from ._check_credentials import create_non_existing_user as create_non_existing_user
@@ -83,6 +47,9 @@ from ._check_credentials import (
 from ._check_credentials import user_exists as user_exists
 from ._check_credentials import user_exists_according_to_profile as user_exists_according_to_profile
 from ._check_credentials import user_locked as user_locked
+from ._connections import locked_attributes as locked_attributes
+from ._connections import multisite_attributes as multisite_attributes
+from ._connections import non_contact_attributes as non_contact_attributes
 from ._need_to_change_pw import is_automation_user as is_automation_user
 from ._need_to_change_pw import need_to_change_pw as need_to_change_pw
 from ._on_access import on_access as on_access
@@ -124,37 +91,3 @@ __all__ = [
 def load_plugins() -> None:
     """Plugin initialization hook (Called by cmk.gui.main_modules.load_plugins())"""
     utils.load_web_plugins("userdb", globals())
-
-
-# The saved configuration for user connections is a bit inconsistent, let's fix
-# this here once and for all.
-def _fix_user_connections() -> None:
-    for cfg in active_config.user_connections:
-        # Although our current configuration always seems to have a 'disabled'
-        # entry, this might not have always been the case.
-        cfg.setdefault("disabled", False)
-        # Only migrated configurations have a 'type' entry, all others are
-        # implictly LDAP connections.
-        cfg.setdefault("type", "ldap")
-
-
-def locked_attributes(connection_id: str | None) -> Sequence[str]:
-    """Returns a list of connection specific locked attributes"""
-    return _get_attributes(connection_id, lambda c: c.locked_attributes())
-
-
-def multisite_attributes(connection_id: str | None) -> Sequence[str]:
-    """Returns a list of connection specific multisite attributes"""
-    return _get_attributes(connection_id, lambda c: c.multisite_attributes())
-
-
-def non_contact_attributes(connection_id: str | None) -> Sequence[str]:
-    """Returns a list of connection specific non contact attributes"""
-    return _get_attributes(connection_id, lambda c: c.non_contact_attributes())
-
-
-def _get_attributes(
-    connection_id: str | None, selector: Callable[[UserConnector], Sequence[str]]
-) -> Sequence[str]:
-    connection = get_connection(connection_id)
-    return selector(connection) if connection else []
