+++ cmk/gui/userdb/_connections.py
@@ -1,0 +1,44 @@
+#!/usr/bin/env python3
+# Copyright (C) 2019 Checkmk GmbH - License: GNU General Public License v2
+# This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
+# conditions defined in the file COPYING, which is part of this source code package.
+
+from collections.abc import Sequence
+from typing import Callable
+
+from cmk.gui.config import active_config
+from cmk.gui.plugins.userdb.utils import get_connection, UserConnector
+
+
+# The saved configuration for user connections is a bit inconsistent, let's fix
+# this here once and for all.
+def fix_user_connections() -> None:
+    for cfg in active_config.user_connections:
+        # Although our current configuration always seems to have a 'disabled'
+        # entry, this might not have always been the case.
+        cfg.setdefault("disabled", False)
+        # Only migrated configurations have a 'type' entry, all others are
+        # implictly LDAP connections.
+        cfg.setdefault("type", "ldap")
+
+
+def locked_attributes(connection_id: str | None) -> Sequence[str]:
+    """Returns a list of connection specific locked attributes"""
+    return _get_attributes(connection_id, lambda c: c.locked_attributes())
+
+
+def multisite_attributes(connection_id: str | None) -> Sequence[str]:
+    """Returns a list of connection specific multisite attributes"""
+    return _get_attributes(connection_id, lambda c: c.multisite_attributes())
+
+
+def non_contact_attributes(connection_id: str | None) -> Sequence[str]:
+    """Returns a list of connection specific non contact attributes"""
+    return _get_attributes(connection_id, lambda c: c.non_contact_attributes())
+
+
+def _get_attributes(
+    connection_id: str | None, selector: Callable[[UserConnector], Sequence[str]]
+) -> Sequence[str]:
+    connection = get_connection(connection_id)
+    return selector(connection) if connection else []
