+++ cmk/gui/userdb/registration.py
@@ -8,10 +8,10 @@ from cmk.gui.config import register_post_config_load_hook
 from cmk.gui.cron import register_job
 from cmk.gui.pages import PageRegistry
 from cmk.gui.plugins.userdb.utils import UserAttributeRegistry, UserConnectorRegistry
-from cmk.gui.userdb import user_attributes
-from cmk.gui.userdb.ldap_connector import register as ldap_connector_register
 
-from . import _fix_user_connections, update_config_based_user_attributes
+from . import ldap_connector, user_attributes
+from ._connections import fix_user_connections
+from ._user_attribute import update_config_based_user_attributes
 from ._user_profile_cleanup import execute_user_profile_cleanup_job, UserProfileCleanupBackgroundJob
 from ._user_sync import ajax_sync, execute_userdb_job, UserSyncBackgroundJob
 
@@ -26,7 +26,7 @@ def register(
 ) -> None:
     user_attributes.register(user_attribute_registry)
 
-    register_post_config_load_hook(_fix_user_connections)
+    register_post_config_load_hook(fix_user_connections)
     register_post_config_load_hook(update_config_based_user_attributes)
     register_job(execute_userdb_job)
 
@@ -36,4 +36,4 @@ def register(
     page_registry.register_page_handler("ajax_userdb_sync", ajax_sync)
     job_registry.register(UserSyncBackgroundJob)
 
-    ldap_connector_register(user_connector_registry)
+    ldap_connector.register(user_connector_registry)
